+++ demos/devkit-simple-tutorial/src/views/Login.js
@@ -1,6 +1,17 @@
-import { LitElement, html, css } from 'lit';
+import { LitElement, css } from 'lit';
 
-import '@blindnet/prci'; // WTF
+import '@blindnet/prci';
+
+import { Auth0Client } from '@auth0/auth0-spa-js';
+
+const auth0 = new Auth0Client({
+  domain: 'blindnet.eu.auth0.com',
+  client_id: '1C0uhFCpzvJAkFi4uqoq2oAWSgQicqHc',
+  redirect_uri: `${window.location.origin}/demos/devkit-simple-tutorial/privacy`,
+  authorizationParams: {
+    redirect_uri: `${window.location.origin}/demos/devkit-simple-tutorial/privacy`,
+  },
+});
 
 export class AppLogin extends LitElement {
   static get properties() {
@@ -12,15 +23,8 @@ export class AppLogin extends LitElement {
   }
 
   render() {
-    return html`
-      <bldn-magic-auth
-        prompt="Please enter your email to submit a request."
-        magic-pub-key="pk_live_7BFA1BB7D19A6FD9"
-        redirect-uri="${window.location
-          .origin}/demos/devkit-simple-tutorial/privacy"
-      >
-      </bldn-magic-auth>
-    `;
+    auth0.getTokenSilently();
+    auth0.loginWithRedirect();
   }
 }