+++ demos/devkit-simple-tutorial/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/blindnet-io/privacy-components-web/compare/v0.9.0...v0.10.0) (2022-11-02)
+
+
+### Bug Fixes
+
+* **@blindnet-demos/devkit-simple-tutorial:** content width ([a61a742](https://github.com/blindnet-io/privacy-components-web/commit/a61a74256e30545affbf05b6a87546f6ccde6e34))
+* **@blindnet/dci:** make DCI requests always have a token ([f6f0add](https://github.com/blindnet-io/privacy-components-web/commit/f6f0add8f2b1dc56c7d323255b484f24f91d62e2))
+
+
+### Features
+
+* **@blindnet-demos/devkit-simple-tutorial:** authentication in demo ([0456da2](https://github.com/blindnet-io/privacy-components-web/commit/0456da2d16148083a9b9c00b09a52f051f0d8fc9))
+* **@blindnet-demos/devkit-simple-tutorial:** demo consent ([#190](https://github.com/blindnet-io/privacy-components-web/issues/190)) ([40e4eec](https://github.com/blindnet-io/privacy-components-web/commit/40e4eec387d3255a36465317f82204c8fefdda70))
+* **@blindnet/core:** magic link authentication component ([08433a6](https://github.com/blindnet-io/privacy-components-web/commit/08433a6f453112f45ab8d98cea861f98bf1314a5))
+* **@blindnet/dci:** new DCI UI ([#159](https://github.com/blindnet-io/privacy-components-web/issues/159)) ([c7b7c8e](https://github.com/blindnet-io/privacy-components-web/commit/c7b7c8ea57802e57a4080452cf4eb565d0899628)), closes [#145](https://github.com/blindnet-io/privacy-components-web/issues/145)
+* **@blindnet/devkit-simple-tutorial:** demo authentication with auth0 ([d61e175](https://github.com/blindnet-io/privacy-components-web/commit/d61e1753961ef5d8bc962855d248be7df20f0bb9))
+
+
+
+
+
+# [0.9.0](https://github.com/blindnet-io/privacy-components-web/compare/v0.8.1...v0.9.0) (2022-09-29)
+
+
+### Bug Fixes
+
+* **@blindnet-demos/devkit-simple-tutorial:** content width ([c5d5cba](https://github.com/blindnet-io/privacy-components-web/commit/c5d5cba5e477533efe492b87aae05ea5d1575a09))
+
+
+### Features
+
+* **@blindnet/dci:** new DCI UI ([#159](https://github.com/blindnet-io/privacy-components-web/issues/159)) ([bf57532](https://github.com/blindnet-io/privacy-components-web/commit/bf57532686c1e5a430dc9f3a53af4c1b217e188e)), closes [#145](https://github.com/blindnet-io/privacy-components-web/issues/145)
+
+
+
+
+
 ## [0.8.1](https://github.com/blindnet-io/privacy-components-web/compare/v0.8.0...v0.8.1) (2022-09-19)
 
 **Note:** Version bump only for package @blindnet-demos/devkit-simple-tutorial